test/web-platform-tests/to-upstream/html/semantics/tabular-data/the-table-element/insertRow-method-03.html

<!DOCTYPE html>
<title>insertRow(): table with a tfoot</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-table-insertrow">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<TABLE ID="table-1" ALIGN="center" BORDER="4" BGCOLOR="#ff0000" FRAME="border" CELLPADDING="2" CELLSPACING="2" SUMMARY="HTML Control Table" RULES="all" WIDTH="680">
<CAPTION ALIGN="top">Table Caption</CAPTION>
<THEAD ALIGN="center" VALIGN="middle">
<TR ALIGN="center" BGCOLOR="#00FFFF" VALIGN="middle">
<TH ID="header-1">Employee Id</TH>
<TH ID="header-2" ABBR="maiden" AXIS="center" ALIGN="center" BGCOLOR="#00FFFF" COLSPAN="1" HEIGHT="50" NOWRAP="nowrap" ROWSPAN="1" SCOPE="col" HEADERS="header-1" VALIGN="middle" WIDTH="100">Employee Name</TH>
<TH>Position</TH>
<TH>Salary</TH>
<TH>Gender</TH>
<TH>Address</TH>
</TR>
</THEAD>
<TFOOT ALIGN="center" VALIGN="middle">
<TR id="last-tr">
<TH>next page ...</TH>
<TH>next page ...</TH>
<TH>next page ...</TH>
<TH>next page ...</TH>
<TH>next page ...</TH>
<TH>next page ...</TH>
</TR>
</TFOOT>
<TBODY ALIGN="center" VALIGN="middle">
<TR>
<TD AXIS="center" ID="Table-3" ABBR="maiden2" ALIGN="center" BGCOLOR="#FF0000" COLSPAN="1" HEIGHT="50" NOWRAP="nowrap" ROWSPAN="1" SCOPE="col" HEADERS="header-2" VALIGN="middle" WIDTH="175">EMP0001</TD>
<TD HEADERS="header-2">Margaret Martin</TD>
<TD>Accountant</TD>
<TD>56,000</TD>
<TD>Female</TD>
<TD>1230 North Ave. Dallas, Texas 98551</TD>
</TR>
<TR>
<TD>EMP0002</TD>
<TD>Martha Raynolds</TD>
<TD>Secretary</TD>
<TD>35,000</TD>
<TD>Female</TD>
<TD>1900 Dallas Road Dallas, Texas 98554</TD>
</TR>
</TBODY>
</TABLE>

<script>
"use strict";

test(() => {

  const table = document.querySelector("table");
  const lastTR = document.querySelector("#last-tr");

  assert_equals(table.rows[table.rows.length - 1], lastTR);

  const tr = table.insertRow(-1);
  assert_equals(tr.parentNode, table.tFoot);
  assert_equals(tr.previousElementSibling, lastTR);

}, "insertRow(-1) should put the element in the tfoot, if it exists (i.e. the parent of the last tr in rows)");

</script>
